// frontend/src/api/model/compositeDashboardSummary.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { CoordinatorKpiSection } from './coordinatorKpiSection';
import type { ExecutiveKpiSection } from './executiveKpiSection';
import type { TechnicianKpiSection } from './technicianKpiSection';

export interface CompositeDashboardSummary {
  userId?: string;
  grantedPermissions?: string[];
  coordinatorSection?: CoordinatorKpiSection;
  technicianSection?: TechnicianKpiSection;
  executiveSection?: ExecutiveKpiSection;
}