// frontend/src/api/model/observationSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface ObservationSummary {
  observationId?: string;
  indicatorId?: string;
  indicatorCode?: string;
  indicatorTitle?: string;
  description?: string;
  issueDate?: string;
  dueDate?: string;
  remainingDays?: number;
  status?: string;
  remediationUrl?: string;
}